requirements.txt
fastapi~=0.115.7
uvicorn~=0.34.0
numpy~=1.26.4
pygame~=2.6.1
rich~=13.9.4
python-docx~=1.1.2
odfpy~=1.4.1
pypdf~=5.2.0
pydantic~=2.10.6
python-multipart~=0.0.20
soundfile~=0.13.1
kokoro~=0.3.3
misaki~=0.6.4
sqlalchemy~=2.0.37
